interface分割
