+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/controller/member/change-progress-status-to-next-controller.ts
@@ -1,9 +1,7 @@
-import {
-  IController,
-  MiddlewareServices,
-  RequestServices,
-  ResponseServices,
-} from "presentation/interface/controller";
+import { IController } from "presentation/interface/controller";
+import { MiddlewareServices } from "presentation/interface/middleware-services";
+import { RequestServices } from "presentation/interface/request-services";
+import { ResponseServices } from "presentation/interface/response-services";
 import * as Usecase from "usecase";
 
 export class ChangeProgressStatusToNextController implements IController {
