+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/controller/member/get-member-controller.ts
@@ -1,9 +1,7 @@
-import {
-  IController,
-  MiddlewareServices,
-  RequestServices,
-  ResponseServices,
-} from "presentation/interface/controller";
+import { IController } from "presentation/interface/controller";
+import { MiddlewareServices } from "presentation/interface/middleware-services";
+import { RequestServices } from "presentation/interface/request-services";
+import { ResponseServices } from "presentation/interface/response-services";
 import { MemberRouteRegistrar } from "presentation/route-registrar";
 import * as Usecase from "usecase";
