+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/implementation/express/controller/controller-service-factory.ts
@@ -1,10 +1,8 @@
 import { NextFunction, Request, Response } from "express";
 import * as Controller from "presentation/implementation/express/controller";
-import {
-  MiddlewareServices,
-  RequestServices,
-  ResponseServices,
-} from "presentation/interface/controller";
+import { MiddlewareServices } from "presentation/interface/middleware-services";
+import { RequestServices } from "presentation/interface/request-services";
+import { ResponseServices } from "presentation/interface/response-services";
 
 type Return = {
   requestServices: RequestServices;
