+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/implementation/express/controller/modules/requesst-services/get-path-params-service.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { IGetPathParams } from "presentation/interface/controller";
+import { IGetPathParams } from "presentation/interface/request-services";
 
 export class GetPathParamsService implements IGetPathParams {
   public constructor(private readonly req: Request) {}
