+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/implementation/express/controller/modules/requesst-services/get-query-params-service.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { IGetQueryParams } from "presentation/interface/controller";
+import { IGetQueryParams } from "presentation/interface/request-services";
 
 export class GetQueryParamsService implements IGetQueryParams {
   public constructor(private readonly req: Request) {}
