+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/implementation/express/controller/modules/requesst-services/get-request-body.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { IGetRequstBody } from "presentation/interface/controller";
+import { IGetRequstBody } from "presentation/interface/request-services";
 
 export class GetRequestBody implements IGetRequstBody {
   public constructor(private readonly req: Request) {}
