+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/implementation/express/controller/modules/response-services/send-response-service.ts
@@ -1,5 +1,5 @@
 import { Response } from "express";
-import { ISendResponse } from "presentation/interface/controller";
+import { ISendResponse } from "presentation/interface/response-services";
 
 export class SendResponseService implements ISendResponse {
   public constructor(private readonly res: Response) {}
