+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/implementation/express/controller/modules/response-services/set-response-status-service.ts
@@ -1,5 +1,5 @@
 import { Response } from "express";
-import { ISetResponseStatus } from "presentation/interface/controller";
+import { ISetResponseStatus } from "presentation/interface/response-services";
 
 export class SetResponseStatusService implements ISetResponseStatus {
   public constructor(private readonly res: Response) {}
