+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/interface/middleware-services.ts
@@ -1,0 +1,7 @@
+export type MiddlewareServices = {
+  nextFunction: INextFunction;
+};
+
+export interface INextFunction {
+  execute: (error?: unknown) => void;
+}
