+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/interface/request-services.ts
@@ -1,0 +1,17 @@
+export type RequestServices = {
+  getPathParams: IGetPathParams;
+  getQueryParams: IGetQueryParams;
+  getRequestBody: IGetRequstBody;
+};
+
+export interface IGetPathParams {
+  execute: () => { [key: string]: string };
+}
+
+export interface IGetQueryParams {
+  execute: () => { [key: string]: string };
+}
+
+export interface IGetRequstBody {
+  execute: () => { [key: string]: unknown };
+}
