+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/presentation/interface/response-services.ts
@@ -1,0 +1,12 @@
+export type ResponseServices = {
+  setResponseStatus: ISetResponseStatus;
+  sendResponse: ISendResponse;
+};
+
+export interface ISetResponseStatus {
+  execute: (statusCode: number) => void;
+}
+
+export interface ISendResponse {
+  execute: (data?: unknown) => void;
+}
